+++ packages/svelte/src/lib/components/nodes/DefaultNode.svelte
@@ -7,36 +7,13 @@
   interface $$Props extends NodeProps {}
 
   export let data: $$Props['data'] = { label: 'Node' };
-  export let targetPosition: $$Props['targetPosition'] = Position.Top;
-  export let sourcePosition: $$Props['sourcePosition'] = Position.Bottom;
+  export let targetPosition: $$Props['targetPosition'] = undefined;
+  export let sourcePosition: $$Props['sourcePosition'] = undefined;
 
-  // unused props - we need to list them here in order to prevent warnings
-  export let id: $$Props['id'] = '';
-  export let width: $$Props['width'] = undefined;
-  export let height: $$Props['height'] = undefined;
-  export let selected: $$Props['selected'] = undefined;
-  export let type: $$Props['type'] = undefined;
-  export let dragging: $$Props['dragging'] = false;
-  export let dragHandle: $$Props['dragHandle'] = undefined;
-  export let positionAbsoluteX: $$Props['positionAbsoluteX'] = 0;
-  export let positionAbsoluteY: $$Props['positionAbsoluteY'] = 0;
-  export let isConnectable: $$Props['isConnectable'];
-  export let zIndex: $$Props['zIndex'];
-
-  // @todo: there must be a better way to do this
-  id;
-  width;
-  height;
-  selected;
-  type;
-  zIndex;
-  dragging;
-  dragHandle;
-  positionAbsoluteX;
-  positionAbsoluteY;
-  isConnectable;
+  // this is a workaround for suppressing the warning about unused props
+  $$restProps;
 </script>
 
-<Handle type="target" position={targetPosition} />
+<Handle type="target" position={targetPosition ?? Position.Top} />
 {data?.label}
-<Handle type="source" position={sourcePosition} />
+<Handle type="source" position={sourcePosition ?? Position.Bottom} />